Deploy commands every launch
Add function to deploy application commands every time you start the bot.
