+++ index.js
@@ -1,71 +1,74 @@
 const readline = require('readline');
 var TokenWizard = require('./public-script/NewToken.js')('.');
+var DeployWizard = require('./public-script/deploy-commands.js')('.');
 
 const rl = readline.createInterface({
 	input: process.stdin,
 	output: process.stdout
-  });
+});
 
 rl.question('Set new token? y/n: ', function (Result) {
-	if(Result === 'y'){
-		TokenWizard.new(function(){
+	if (Result === 'y') {
+		TokenWizard.new(function () {
 			console.log('Completed token wizard!');
 			StartBot();
-		  });
+		});
 	}
-	if(Result === 'n'){
+	if (Result === 'n') {
 		StartBot();
-	}
-  });
-
-//Base bot functionality
-function StartBot(){
-	console.log('Starting bot!')
-	const fs = require('fs');
-const { Client, Collection, Intents } = require('discord.js');
-const { token } = require('./config.json');
-
-const client = new Client({ intents: [Intents.FLAGS.GUILDS] });
-
-const eventFiles = fs.readdirSync('./events').filter(file => file.endsWith('.js'));
-
-for (const file of eventFiles) {
-	const event = require(`./events/${file}`);
-	if (event.once) {
-		client.once(event.name, (...args) => event.execute(...args));
-	}
-	else {
-		client.on(event.name, (...args) => event.execute(...args));
-	}
-}
-
-client.commands = new Collection();
-
-const commandFiles = fs.readdirSync('./commands').filter(file => file.endsWith('.js'));
-
-for (const file of commandFiles) {
-	const command = require(`./commands/${file}`);
-	// Set a new item in the Collection
-	// With the key as the command name and the value as the exported module
-	client.commands.set(command.data.name, command);
-}
-
-client.on('interactionCreate', async interaction => {
-	if (!interaction.isCommand()) return;
-
-	const command = client.commands.get(interaction.commandName);
-
-	if (!command) return;
-
-	try {
-		await command.execute(interaction);
-	}
-	catch (error) {
-		console.error(error);
-		await interaction.reply({ content: 'There was an error while executing this command!', ephemeral: true });
 	}
 });
 
-client.login(token);
-	
+//Base bot functionality
+function StartBot() {
+	DeployWizard.deploy(function () {
+		console.log('Successfully registered application commands!');
+		console.log('Starting bot...');
+		const fs = require('fs');
+		const { Client, Collection, Intents } = require('discord.js');
+		const { token } = require('./config.json');
+
+		const client = new Client({ intents: [Intents.FLAGS.GUILDS] });
+
+		const eventFiles = fs.readdirSync('./events').filter(file => file.endsWith('.js'));
+
+		for (const file of eventFiles) {
+			const event = require(`./events/${file}`);
+			if (event.once) {
+				client.once(event.name, (...args) => event.execute(...args));
+			}
+			else {
+				client.on(event.name, (...args) => event.execute(...args));
+			}
+		}
+
+		client.commands = new Collection();
+
+		const commandFiles = fs.readdirSync('./commands').filter(file => file.endsWith('.js'));
+
+		for (const file of commandFiles) {
+			const command = require(`./commands/${file}`);
+			// Set a new item in the Collection
+			// With the key as the command name and the value as the exported module
+			client.commands.set(command.data.name, command);
+		}
+
+		client.on('interactionCreate', async interaction => {
+			if (!interaction.isCommand()) return;
+
+			const command = client.commands.get(interaction.commandName);
+
+			if (!command) return;
+
+			try {
+				await command.execute(interaction);
+			}
+			catch (error) {
+				console.error(error);
+				await interaction.reply({ content: 'There was an error while executing this command!', ephemeral: true });
+			}
+		});
+
+		client.login(token);
+	});
 }
